API/product-service/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-DotEnv==0.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
werkzeug==3.0.1
pyyaml